Cache list of targets.
Getting the list of targets can be extremely slow. Since the targets
don't change often, we can cache them. The first call will be slow,
however all further call will be extremely fast, since we read from a
file.

We rely on the modified time of the BUILD file to decide that our
cache is out of date.